sim_data_input_dashboard.R
This script creates produces a dummy dataset of daily population estimates by oblast, sex and age groups to be used as input for the dashboard. Data is in long format. Dataset has 10 columns and 654160 rows. Columns correspond to admin1 codes (oblasts), sex, age groups, date of the day, mean, median, upper and lower of 60% and 95% confidence intervals for population estimates. The number of rows correspond to 
-	26 geographical units (oblasts) plus rest of the world
-	2 sexes (female, male)
-	17 five-year age groups (from 0 to 85+)
-	740 days counting from 24 February 2022 until today (4 March 2024)