code for model on proportions
Uploading the code for the model on proportions. There are 2 versions: RW (random walk on the logit-scale proportion parameters); RE (time and location random effects on these parameters, akin to the random effects in the Facebook proportion model). Both models appear to fit the data well. My next task will be to run model selection, comparing these two via LOO-CV to see if one is substantially better than the other.